+++ app/src/main/java/com/zaki/movieapp/data/remote/response/MovieTrendingResponse.kt
@@ -12,5 +12,5 @@ data class MovieTrending(
     @SerializedName("overview") val overview: String? = null,
     @SerializedName("release_date") val releaseDate: String? = null,
     @SerializedName("poster_path") val posterPath: String? = null,
-    @SerializedName("vote_average") val voteAverage: Int? = null
+    @SerializedName("vote_average") val voteAverage: Double? = null
 )
